Fix tests that fail when DISPLAY_UNITS=mmol
Some of the tests check for values in units of mg/dL.  These fail when
the `DISPLAY_UNITS` environment variable is set to `mmol` or `mmol/L`.

The failures can be reproduced with:

```
$ DISPLAY_UNITS=mmol npm run-script test
$ DISPLAY_UNITS=mmol/L npm run-script test
```

This change fixes the failing cases by first checking the configured
units, and then choosing an expected value accordingly.

This fix was tested with:

```
$ npm run-script test
$ DISPLAY_UNITS=mg/dl npm run-script test
$ DISPLAY_UNITS=mmol npm run-script test
$ DISPLAY_UNITS=mmol/L npm run-script test
```

Fixes #8353